+++ src/ts/draw.ts
@@ -42,7 +42,9 @@
  * @modified 2022-02-03 Added the `lineWidth` param to the `crosshair` function.
  * @modified 2022-02-03 Added the `cross(...)` function.
  * @modified 2022-03-27 Added the `texturedPoly` function.
- * @version  1.12.0
+ * @modified 2022-06-01 Tweaked the `polyline` function; lineWidth now scales with scale.x.
+ * @modified 2022-07-26 Adding `alpha` to the `image(...)` function.
+ * @version  1.12.2
  **/
 
 import { CubicBezierCurve } from "./CubicBezierCurve";
@@ -229,16 +231,18 @@ export class drawutils implements DrawLib<void> {
    * @param {Image} image - The image object to draw.
    * @param {Vertex} position - The position to draw the the upper left corner at.
    * @param {Vertex} size - The x/y-size to draw the image with.
+   * @param {number=0.0} alpha - (optional, default=0.0) The transparency (1.0=opaque, 0.0=transparent).
    * @return {void}
    * @instance
    * @memberof drawutils
    **/
-  image(image: HTMLImageElement, position: Vertex, size: Vertex): void {
+  image(image: HTMLImageElement, position: Vertex, size: Vertex, alpha: number = 1.0): void {
     if (!image.complete || !image.naturalWidth) {
       // Avoid drawing un-unloaded or broken images
       return;
     }
     this.ctx.save();
+    this.ctx.globalAlpha = alpha;
     // Note that there is a Safari bug with the 3 or 5 params variant.
     // Only the 9-param varaint works.
     this.ctx.drawImage(
@@ -1003,10 +1007,12 @@ export class drawutils implements DrawLib<void> {
    * @memberof drawutils
    */
   polyline(vertices: Array<Vertex>, isOpen: boolean, color: string, lineWidth?: number) {
-    if (vertices.length <= 1) return;
+    if (vertices.length <= 1) {
+      return;
+    }
     this.ctx.save();
     this.ctx.beginPath();
-    this.ctx.lineWidth = lineWidth || 1.0;
+    this.ctx.lineWidth = lineWidth * this.scale.x || 1.0;
     this.ctx.moveTo(this.offset.x + vertices[0].x * this.scale.x, this.offset.y + vertices[0].y * this.scale.y);
     for (var i = 0; i < vertices.length; i++) {
       this.ctx.lineTo(this.offset.x + vertices[i].x * this.scale.x, this.offset.y + vertices[i].y * this.scale.y);